add moderate user service

+++ src/ChatApp/Services/user-service/src/main/java/com/chatapp/userservice/entity/keys/MessageStatusId.java
@@ -5,6 +5,8 @@ import jakarta.persistence.Embeddable;
 import lombok.*;
 import lombok.experimental.FieldDefaults;
 
+import java.io.Serializable;
+
 @Getter
 @Setter
 @Embeddable
@@ -13,8 +15,8 @@ import lombok.experimental.FieldDefaults;
 @FieldDefaults(level = AccessLevel.PRIVATE)
 public class MessageStatusId {
     @Column(name = "message_id")
-    Long messageId;
+    int messageId;
 
     @Column(name = "user_id")
-    Long userId;
+    int userId;
 }
